+++ packages/math/src/decimal.ts
@@ -127,6 +127,19 @@ export class Decimal {
   }
 
   /**
+   * a.minus(b) returns a-b.
+   *
+   * Both values need to have the same fractional digits.
+   * The resulting difference needs to be non-negative.
+   */
+  public minus(b: Decimal): Decimal {
+    if (this.fractionalDigits !== b.fractionalDigits) throw new Error("Fractional digits do not match");
+    const difference = this.data.atomics.sub(new BN(b.atomics));
+    if (difference.ltn(0)) throw new Error("Difference must not be negative");
+    return new Decimal(difference.toString(), this.fractionalDigits);
+  }
+
+  /**
    * a.multiply(b) returns a*b.
    *
    * We only allow multiplication by unsigned integers to avoid rounding errors.